Add a new PrettyTable parsing function
The current state of (at least some) OpenStack CLI tools do not provide a
mechanism for outputting data in formats other than PrettyTable output.
Therefore this function is intended to parse PrettyTable output into a
usable array of hashes. Similarly, it will flatten Property/Value tables
into a single element array.

This change adds a parser as well as the associated rspec tests for that
parser. The initial use case for this function is to allow the Identity
LWRP to replace HTTP/S calls with Identity CLI calls. As much of what
happens with the CLI involves chaining UUIDs, this function is critical
to make that happen.

Note well: bumping the version of this cookbook to 0.2.1

All tests pass.

Change-Id: Icac7fcd1670f3240d59b07fce25cf8283ba2e0d8