Allow for memcache_servers override attributes
Right now the memcache_servers function is used throughout a number of
the cookbooks. For those who have not defined a memcache role, this
can be problematic as it uses search. This would also be problematic
for those wanting to use chef-solo without search and/or test-kitchen.

So, just provide an override attribute to override search for
memcache_servers.

New test added to support this and all tests pass.

Change-Id: Iff655ff066d8aa933eedb67ac17611288b57a2f0