Revert "Correct faulty Chef search query"
This reverts commit 8311869e5b99fecefd567ce3f1ad1cbdf8d5c5c6.

When a role is present on the current node either the node attributes or a
subset are returned depending if section is passed.  Conversely if role is
not present on the current node search_for is called which always returns an
array.  This is inconsistant.  Since it appears the roles should contain the
attributes it doesn't matter which one is return, so the first should suffice.
Update spec tests stub to properly return an array from search_for

Change-Id: I180fe8c6a4d0e31e877b91f8cbad54018ef81049