Remove config_by_role in favor of search_for
config_by_role is only used in once place:
openstack-compute::nova-common. And, in this single place it happens
to return inconsistent results. Now that we have search_for, which
does return consistent results, move this one last config_by_role
call to use that. In the interest of good clean living, I am removing
this code.

Bumping to version 0.4.0

Change-Id: Ib755a5d568008599ee7d663f43716e7fe47251f9