allow overriding all endpoint hosts with a single attribute
similar to the pattern in the messaging and database attributes, this is
a convenience change that allows you to set all endpoint host values to
something other than the default of 127.0.0.1, by only overriding a
single attribute in your environment.

Quite often one will want to set all endpoints to the same IP, for
example in the case where you are deploying a 'controller' node, and at
the moment every endpoint host attribute has to be overridden
individually.

All endpoint host attributes can still be overridden individually if
desired.

Change-Id: I7dd96f69c3eaa9156513663f08dafbddf1c1ffd8