Fix unnecessary error messages in log
Currently if you are working in an environment that cannot reach the
target openstack-ppa uri/components,say you do not have internet access
for instance, you will see an error in the log which is captured in
the refferenced bug report.

By adding a new flag attribute:
['openstack']['apt']['live_updates_enabled'], the opnstack ppa update
can be disabled by setting this attribute to false.

This patch intended to avoid executing an unnesessary code path by
checking the new flag  prior to attempting to fetch updates.

Closes-Bug: #1289356
Change-Id: Ifca7c77e7a84834b88600c14d5a45cc0a64825f8