Add python-openstackclient package support
Add new client recipe for new common client package.
As more new features are only supported via this new python client, we
need to have a way to deploy it and Common seems like the right place.
There is another blueprint for moving the openrc from Compute to Common
that is related to this.

Change-Id: I8ab56b375b28c979021f3c5bc81e364f1cc65226
Implements: blueprint openstack-client