Add [<service>-bind] attributes to each service namespace
Similar to what already existed for identity, this change adds a pair of *-bind
attributes for each service.  This allows for specifying a different
bind host or interface to actually bind a service to on each host, versus the
endpoint for the service which may be a separate IP on a load balancer.

This change is a pre-requisite change needed for the subsequent work in
each cookbook.

Change-Id: I3b0ee00a72bbe786b4ed399f79bb3c6f4c46167f