Add missing AMQP attributes
There are some settings that are useful for tweaking AMQP
settings in all the service configurations that should
be made available for setting. This adds flags for durable_queues,
auto_delete, and a setting for QPID topology version that
is required to be set correctly to use QPID federation.

There will need to be dependent changes to utilize these
in the project specific cookbooks.

Change-Id: I850c01a40aab02cf4531f7983543ef89e0e60446