Update default['openstack']['yum']['repo-key']
Some yum commands are failing as raw.github.com now redirects to
raw.githubusercontent.com.  This change simply updates to the new
destination URL to avoid the redirect.

Additionally, I have to bump fauxhai to >= 2.1 as fauxhai tests are
failing due to this same issue. We're using an older version of fauxhai
with no support for Red Hat 6.5, so it tries to connect out to GitHub
to pull down platform/version JSON.

Change-Id: I33d6b63ecbb8f0698b0b8f6d92ef93c279ff4916