allow for attribute storage of secrets as an alternative
In addition to storing secrets in data bags (the default) we would like
to be able to store them in attributes.

This commit doesn't change the existing passwords API, get_password will
now read from attributes when the node['openstack']['use_databags']
attribute is falsy.

This commit deprecates the development_mode which was only used as a
hack for getting passwords easier without data bags. The attribute
storage of passwords should now be sufficient for all those use cases.

Increased rubocop's methodlength maximum to 15 because there are some
temporary deprecation warnings which add up, but should go away soon.

Change-Id: I1f23878dd3fa83fb40f7b4b56960d57a7b9b89cc