Attribute variables for internal identity endpoint
This patch adds the identity-internal-endpoints attributes to complment
the first patch of this bug.  As cookbook-openstack-identity now
refences the  identity-internal endpoint these variables are now
required to be defined.  As per convention endpoints are defined in this
cookbook.

Following this patch the number of node.set used in the registration
tests can be reduced, as they will now have defaults defined here.
Additionally a bump of the identity cookbook to rely on common 10.0.3
will be required.
Once this is done this bug can be closed.

Change-Id: Iefca0b3df5b8efae94a661fc471c55a05cd8f64e
Partial-Bug: #1356732