Revert the default cinder api version to v1
Since there are many problems when applying cinder v2 api, like
cannot attach volumes to VM from nova, and there may be other
problems in glance, the integration of cinder v2 api and
other services is not well tested and the related configurations
are not clear yet. We'd better revert the cinder api version back
to v1 before all things are clear.

Closes-bug: #1388741
Change-Id: Ic27a437be3c45e8a522c9c50b1e658de4ee79e94