Removed the hardcoded PPA reference to "Precise"
For Ubuntu installations greater than version 12.04 a PPA was being
inserted with a reference to precise. This is problomatic due to
Juno not being backported into Precise.  When the PPA is inserted
as is, dpkg has errors.

The fix was to use the codename information from ohai in the default
attributes to determine which PPA to use.

Changelog update in tow

Change-Id: I2eba21c5384cb50f32686d9de080613c1af7f203