specific_endpoint can return incorrect values
There's a bug in the common library specific_endpoint routine that
was just merged. If there's a defined endpoint for any one of the
specified types ('internal', 'admin', 'public') it causes the code
to incorrectly think that any name of that same type actually
exists.  This leads to an empty string being returned rather than
picking up the general endpoint.

For example:
['endpoints']['admin']['foo'] has been set
['endpoints']['bar'] has been set

The admin_endpoint('bar') routine should see that there's no 'bar'
under 'admin' and return the general endpoint. Instead, it's
returning an empty string simply because there's an 'admin' subhash
there for 'foo'.

Change-Id: Ifd4ee6108b4c7b0f66bd56cc817b5ea646e82a68
Partial-Bug: 1412919