Add identity_uri_transform to help move to identity_uri
Starting on the identity_uri blueprint.  A difference
between the public auth_uri and the admin identity_uri is that
the identity_uri does not include the path/version (/v2.0).
And the auth_uri has to modify the version depending upon
the needs of the environment.
There existed a helper routine called, auth_uri_transform, so
to be consistent and make this easy to follow, I created a
identity_uri_transform which simply removes the path/version
from the uri given.

This helper will then be used thru out the cookbooks to make
the switch to the identity_uri very easy.

I guess the alturnatives are messing with our generic endpoint
definitions/methods, but since they are used for many different
purposes, I think having another matching transform method is the
right way to go here.

Bumped version for adding new method which other cookbooks will
depend upon.

Change-Id: I035a4491945c600dea8feb317b357a1a69da28cd
Implements: blueprint identity-uri