Fix render_config_file with_section_content to handle dup sections
Since duplicate sections are allowed within openstack conf files,
need to be able to handle them in some cases.  The first case of
this is in network where the ml2 conf is including the openvswitch
conf within it and there are a couple dup sections.  This
allows the spec tests to handle those cases.

Change-Id: I8d2733a248a6a13ab87c286ff620c71955726f86
Closes-Bug: #1444696