Revert "update the apt cookbook version".
The update of the apt cookbook version currently breaks the gates for opentstack-chef-repo and openstack-block-storage, because of a dependency on the ceph cookbook, which is incompatible in its currently released version with the new apt cookbook. The fix was committed in master (https://github.com/ceph/ceph-cookbook/commit/4883a0b09232e2cc205851709b80b168e7ee5450) and we can do our apt update as soon, as this is released to the supermarket.

This reverts commit f0e7724f37516ed39eaf17adb2e3f2273347c162.

Change-Id: Ib36b58e7e25c4c24bb480fc9412d909c167f5fdd