CI - Depends-On support
This patch will use the Depends-On info stored in the ZUUL_CHANGES
env var to update the Berksfile to a local patch and then use
zuul-cloner to pull down that dependency.

To use Depends-On, add the following to your commit message AFTER
the description and BEFORE the change id and bug/blueprint ref.
"Depends-On: <id>"
The id is the change id of the gerrit patch this patch depends on.

Change-Id: I13a5ef21733cdd00841dc1ec1afe7442b2dc8831