Add recipe to create bash completions for OSC
The new recipe "completions" can be executed in order to create a file
with command completions for the "openstack" command in bash. We want
this to be dynamic because the set of available commands depends on the
set of python plugins installed on the host.

Bump minor version for stable branch.

Change-Id: Ib6549c00596f6560dab74491309b6b580ae7b312
(cherry picked from commit 9170a08c7c4c427438d4353ff82438376d3ea3ac)