Drop poise-python in favor of directly installing python packages
The poise-python cookbook is no longer maintained by the upstream maintainer and
is causing lots of issues for the testing pipeline. Originally we needed this to
install a newer openstack-client via pip, but we no longer do that.

The one exception is in the integration cookbook which I'll be pushing a change
for that as well.

In order to do this, we need to include EPEL on RHEL which doesn't seem to have
any current dependency issues so I think it's safe to bring back.

Depends-On: https://review.opendev.org/665493
Change-Id: Ic8ddf736193a6a7cea5372f2fd09e2def3634333
Signed-off-by: Lance Albertson <lance@osuosl.org>